pyproject.toml
[project]
name = "unraveled"
version = "0.1.0"
description = "Reverse engineering agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "floki-ai>=0.7.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "uvicorn>=0.34.0",
    "durabletask>=0.1.1a1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unraveled"]

[dependency-groups]
dev = ["pytest>=8.3.4"]

[tool.hatch.build]
packages = ["src/unraveled"]
